% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{bc}
\alias{bc}
\title{An example back check of an example survey}
\format{A data frame with 14 rows and 6 variables:
\describe{
  \item{id}{Unique ID}
  \item{date}{Back check date}
  \item{bcer}{Back checker name}
  \item{gender}{Respondent gender}
  \item{gameresult}{Game result}
  \item{itemssold}{Number of items sold}
}}
\source{
\url{https://github.com/PovertyAction/bcstats}
}
\usage{
bc
}
\description{
A dataset containing an exapmle survey from an example back check
}
\keyword{datasets}

