% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{survey}
\alias{survey}
\title{An example survey}
\format{A data frame with 33 rows and 7 variables:
\describe{
  \item{id}{Unique ID}
  \item{date}{Interview date}
  \item{enum}{Enumerator}
  \item{enumteam}{Enumerator team}
  \item{gender}{Respondent gender}
  \item{gameresult}{Game result}
  \item{itemssold}{Number of items sold}
}}
\source{
\url{https://github.com/PovertyAction/bcstats}
}
\usage{
survey
}
\description{
A dataset containing an example survey
}
\keyword{datasets}

