% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcstats.r
\name{bcstats}
\alias{bcstats}
\title{Comparing "back checks" in R (a clone of Stata's bcstats)}
\usage{
bcstats(surveydata, bcdata, id, t1vars = NA, t2vars = NA, t3vars = NA)
}
\arguments{
\item{surveydata}{The survey data}

\item{bcdata}{The back check data}

\item{id}{the unique ID}

\item{t1vars}{The list of "type 1" variables}

\item{t2vars}{The list of "type 2" variables}

\item{t3vars}{The list of "type 3" variables}
}
\value{
A data.frame with the error rates
}
\description{
Comparing "back checks" in R (a clone of Stata's bcstats)
}

